+++ test/pagination.spec.js
@@ -5,9 +5,24 @@ const { getQueries } = require('./utils');
 const pagination = require('../src/pagination');
 
 const akaMock = (id, query) => `/?${qs.stringify(query)}`.replace(/\?$/, '');
+const requestObjMock = {
+  headers: {},
+  connection: {
+    info: {
+      protocol: 'http',
+    }
+  },
+  info: {
+    host: 'localhost:1337',
+  },
+  url: {
+    pathname: '/'
+  }
+};
 
 test('bissle/pagination.getLinks >> has no prev HAL link | first page', t => {
-  const result = pagination.getLinks('foo', 1, 3, 9, { per_page: 1 }, {}, akaMock);
+  const options = { per_page: 1 };
+  const result = pagination.getLinks('foo', 1, 3, 9, requestObjMock, akaMock, options, {});
 
   t.equal(_.has(result, 'first'), true);
   t.equal(_.has(result, 'prev'), false);
@@ -17,7 +32,8 @@ test('bissle/pagination.getLinks >> has no prev HAL link | first page', t => {
 });
 
 test('bissle/pagination.getLinks >> has no next HAL link | last page', t => {
-  const result = pagination.getLinks('foo', 3, 3, 9, { per_page: 1 }, {}, akaMock);
+  const options = { per_page: 1 };
+  const result = pagination.getLinks('foo', 3, 3, 9, requestObjMock, akaMock, options, {});
 
   t.equal(_.has(result, 'first'), true);
   t.equal(_.has(result, 'prev'), true);
@@ -27,7 +43,8 @@ test('bissle/pagination.getLinks >> has no next HAL link | last page', t => {
 });
 
 test('bissle/pagination.getLinks >> has all HAL links | middle page', t => {
-  const result = pagination.getLinks('foo', 2, 3, 9, { per_page: 1 }, {}, akaMock);
+  const options = { per_page: 1 };
+  const result = pagination.getLinks('foo', 2, 3, 9, requestObjMock, akaMock, options, {});
 
   t.equal(_.has(result, 'first'), true);
   t.deepEqual(getQueries(result.first), {
@@ -55,13 +72,14 @@ test('bissle/pagination.getLinks >> has all HAL links | middle page', t => {
 });
 
 test('bissle/pagination.getLinks >> has all HAL links | customized middle page ', t => {
-  const result = pagination.getLinks('foo', 2, 3, 9, { per_page: 3 }, {}, akaMock);
+  const options = { per_page: 3 };
+  const result = pagination.getLinks('foo', 2, 3, 9, requestObjMock, akaMock, options, {});
 
   t.equal(_.has(result, 'first'), true);
-  t.equal(getQueries(result.first), undefined);
+  t.equal(getQueries(result.first), undefined, 'first');
 
   t.equal(_.has(result, 'prev'), true);
-  t.equal(getQueries(result.prev), undefined);
+  t.equal(getQueries(result.prev), undefined, 'prev');
 
   t.equal(_.has(result, 'next'), true);
   t.deepEqual(getQueries(result.next), {
@@ -77,7 +95,8 @@ test('bissle/pagination.getLinks >> has all HAL links | customized middle page '
 });
 
 test('bissle/pagination.getLinks >> has no prev/next HAL links | page is zero', t => {
-  const result = pagination.getLinks('foo', 0, 3, 9, { per_page: 1 }, {}, akaMock);
+  const options = { per_page: 1 };
+  const result = pagination.getLinks('foo', 0, 3, 9, requestObjMock, akaMock, options, {});
 
   t.equal(_.has(result, 'first'), true);
   t.equal(_.has(result, 'prev'), false);
@@ -88,31 +107,12 @@ test('bissle/pagination.getLinks >> has no prev/next HAL links | page is zero', 
 
 
 test('bissle/pagination.getLinks >> has no prev/next HAL links | page is too high', t => {
-  const result = pagination.getLinks('foo', 4, 3, 9, { per_page: 1 }, {}, akaMock);
+  const options = { per_page: 1 };
+  const result = pagination.getLinks('foo', 4, 3, 9, requestObjMock, akaMock, options, {});
 
   t.equal(_.has(result, 'first'), true);
   t.equal(_.has(result, 'prev'), false);
   t.equal(_.has(result, 'next'), false);
   t.equal(_.has(result, 'last'), true);
-  t.end();
-});
-
-test('bissle/pagination.optimizeLinks >> converts links into relative ones', t => {
-  const links = { next: 'http://localhost:1337/?page=2' };
-  const pluginOptions = { absolute: false };
-
-  pagination.optimizeLinks(links, pluginOptions);
-
-  t.equal(_.some(links, link => _.startsWith(link, 'http://localhost:1337/')), false);
-  t.end();
-});
-
-test('bissle/pagination.optimizeLinks >> does not convert links into relative ones', t => {
-  const links = { next: 'http://localhost:1337/?page=2' };
-  const pluginOptions = { absolute: true };
-
-  pagination.optimizeLinks(links, pluginOptions);
-
-  t.equal(_.every(links, link => _.startsWith(link, 'http://localhost:1337/')), true);
   t.end();
 });